@@ -7,17 +7,14 @@ module.exports = {
   parserOptions: {
     parser: 'babel-eslint'
   },
-  extends: [
-    // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
-    // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
-    'plugin:vue/recommended',
-    '@vue/prettier'
-  ],
+  extends: ['plugin:vue/recommended'],
   // required to lint *.vue files
   plugins: ['vue'],
   // add your custom rules here
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 2 : 0,
-    'no-unused-vars': process.env.NODE_ENV === 'production' ? 2 : 0
+    semi: [2, 'never'],
+    'no-console': 'off',
+    'vue/no-v-html': 'off',
+    'vue/max-attributes-per-line': 'off'
   }
 }
