@@ -1,27 +1,20 @@
 <template>
-  <div
-    :id="name"
-    class="card card-plain">
-    <div
-      class="card-header">
+  <div :id="name" class="card card-plain">
+    <div class="card-header">
       <a
         :class="showCollapse ? 'collapsed' : null"
         :aria-expanded="showCollapse ? 'true' : 'false'"
         :aria-controls="name"
         :href="`#${name}`"
         data-toggle="collapse"
-        @click.prevent="showCollapse = !showCollapse">
+        @click.prevent="showCollapse = !showCollapse"
+      >
         {{ title }}
-        <i class="now-ui-icons arrows-1_minimal-down"/>
+        <i class="now-ui-icons arrows-1_minimal-down" />
       </a>
     </div>
-    <b-collapse
-      :id="name"
-      :visible="visible"
-      v-model="showCollapse">
-      <div
-        class="card-body"
-        v-html="desc"/>
+    <b-collapse :id="name" v-model="showCollapse" :visible="visible">
+      <div class="card-body" v-html="desc" />
     </b-collapse>
   </div>
 </template>
