@@ -1,20 +1,15 @@
-
 <template>
   <div class="wrapper">
     <q-header-small class-name="ben-white">
-      <h1
-        class="title"
-        v-html="article.title"/>
-      <h3 v-html="article.subtitle"/>
+      <h1 class="title" v-html="article.title" />
+      <h3 v-html="article.subtitle" />
     </q-header-small>
     <div class="section">
       <div class="">
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-              <h3
-                class="title"
-                v-html="article.title"/>
+              <h3 class="title" v-html="article.title" />
               <p v-html="article.introduction" />
             </div>
           </div>
@@ -26,28 +21,24 @@
             <div class="row">
               <div class="col-md-8 ml-auto mr-auto">
                 <h3>Where do you start?</h3>
-                <p>To share a single photo you can find them through your <a
-                  href="https://app.skedr.io/photostream"
-                  target="_blank">Photostream</a>* or viewing all images from one of your <a
-                    href="https://app.skedr.io/groups"
-                    target="_blank">Groups</a>* ( and clicking the arrow button). In both cases you will see that each image has a Sked button.</p>
+                <p>
+                  To share a single photo you can find them through your
+                  <a href="https://app.skedr.io/photostream" target="_blank">Photostream</a>* or viewing all images from
+                  one of your <a href="https://app.skedr.io/groups" target="_blank">Groups</a>* ( and clicking the arrow
+                  button). In both cases you will see that each image has a Sked button.
+                </p>
                 <p><i>*You need to be logged into Skedr account</i></p>
                 <br>
-                <div
-                  v-b-modal.sked-button
-                  class="image-container sked-button"/>
-                <b-modal
-                  id="sked-button"
-                  size="lg"
-                  hide-header
-                  hide-footer>
-                  <img
-                    class="image-fluid"
-                    src="~/assets/img/sked-button.png">
+                <div v-b-modal.sked-button class="image-container sked-button" />
+                <b-modal id="sked-button" size="lg" hide-header hide-footer>
+                  <img class="image-fluid" src="~assets/img/sked-button.png">
                 </b-modal>
                 <br>
                 <br>
-                <p>Let's dig into the Share functionality by pressing the Sked button of any of your photos. A popup menu will open with all your groups.</p>
+                <p>
+                  Let's dig into the Share functionality by pressing the Sked button of any of your photos. A popup menu
+                  will open with all your groups.
+                </p>
               </div>
             </div>
           </div>
@@ -63,20 +54,15 @@
                 <ul>
                   <li><strong>Title:</strong> Name of the photo you want to share.</li>
                   <li><strong>Search bar:</strong> To filter through your groups and find the correct one.</li>
-                  <li><strong>Group list:</strong> Each group from the list has it's name, photo ( with a link to that groups page ), members count, pool count and posting limits for that group.</li>
+                  <li>
+                    <strong>Group list:</strong> Each group from the list has it's name, photo ( with a link to that
+                    groups page ), members count, pool count and posting limits for that group.
+                  </li>
                   <li><strong>Bottom toolbar:</strong> A list of selected groups and the send button.</li>
                 </ul>
-                <div
-                  v-b-modal.sked-dialog
-                  class="image-container sked-dialog"/>
-                <b-modal
-                  id="sked-dialog"
-                  size="lg"
-                  hide-header
-                  hide-footer>
-                  <img
-                    class="image-fluid"
-                    src="~/assets/img/sked-dialog.png">
+                <div v-b-modal.sked-dialog class="image-container sked-dialog" />
+                <b-modal id="sked-dialog" size="lg" hide-header hide-footer>
+                  <img class="image-fluid" src="~assets/img/sked-dialog.png">
                 </b-modal>
               </div>
             </div>
@@ -89,22 +75,20 @@
             <div class="row">
               <div class="col-md-8 ml-auto mr-auto">
                 <h3>Choosing which groups you will share your photo</h3>
-                <p>Choose into which groups do you want to share the photo and click the Share button at the bottom toolbar of the dialog menu.</p>
-                <div
-                  v-b-modal.sked-sharing-photos
-                  class="image-container sked-sharing-photos"/>
-                <b-modal
-                  id="sked-sharing-photos"
-                  size="lg"
-                  hide-header
-                  hide-footer>
-                  <img
-                    class="image-fluid"
-                    src="~/assets/img/sked-sharing-photos.png">
+                <p>
+                  Choose into which groups do you want to share the photo and click the Share button at the bottom
+                  toolbar of the dialog menu.
+                </p>
+                <div v-b-modal.sked-sharing-photos class="image-container sked-sharing-photos" />
+                <b-modal id="sked-sharing-photos" size="lg" hide-header hide-footer>
+                  <img class="image-fluid" src="~assets/img/sked-sharing-photos.png">
                 </b-modal>
                 <br>
                 <br>
-                <p>If you still have doubts about this, write to <a href="mailto:info@skedr.io">info@skedr.io</a> for  more help.</p>
+                <p>
+                  If you still have doubts about this, write to <a href="mailto:info@skedr.io">info@skedr.io</a> for
+                  more help.
+                </p>
               </div>
             </div>
           </div>
@@ -120,16 +104,20 @@
                   <div class="col-md-2">
                     <div class="card-avatar">
                       <a href="#pablo">
-                        <img
-                          class="img img-raised"
-                          src="~/assets/img/enric-bisbe-gil.jpg">
+                        <img class="img img-raised" src="~assets/img/enric-bisbe-gil.jpg">
                       </a>
-                      <div class="ripple-container"/>
+                      <div class="ripple-container" />
                     </div>
                   </div>
                   <div class="col-md-8">
-                    <h4 class="card-title">Enric Bisbe Gil</h4>
-                    <p class="description"><a href="https://github.com/ebisbe">Developer</a> and <a href="https://www.flickr.com/photos/enric-bisbe">Photography enthusiast</a>. Creator of Skedr.io with the porpuse of helping others. </p>
+                    <h4 class="card-title">
+                      Enric Bisbe Gil
+                    </h4>
+                    <p class="description">
+                      <a href="https://github.com/ebisbe">Developer</a> and
+                      <a href="https://www.flickr.com/photos/enric-bisbe">Photography enthusiast</a>. Creator of
+                      Skedr.io with the porpuse of helping others.
+                    </p>
                   </div>
                   <!-- <div class="col-md-2">
                     <button
@@ -158,7 +146,7 @@
                         <div class="card-image">
                           <img
                             class="img img-raised rounded"
-                            src="~/assets/img/View_photostream_desktop_small.png">
+                            src="~assets/img/View_photostream_desktop_small.png">
                         </div>
                       </div>
                       <div class="col-md-7">
@@ -202,7 +190,7 @@
                           <img
                             class="img img-raised rounded
                                             "
-                            src="~/assets/img/View_photostream_desktop_small.png">
+                            src="~assetsimg/View_photostream_desktop_small.png">
                         </div>
                       </div>
                     </div>
@@ -246,16 +234,16 @@ export default {
 
 <style>
 .ben-white {
-  background-image: url(~/assets/img/ben-white-226176-unsplash.jpg);
+  background-image: url(~assets/img/ben-white-226176-unsplash.jpg);
 }
 .sked-button {
-  background-image: url(~/assets/img/sked-button.png);
+  background-image: url(~assets/img/sked-button.png);
 }
 .sked-dialog {
-  background-image: url(~/assets/img/sked-dialog.png);
+  background-image: url(~assets/img/sked-dialog.png);
 }
 .sked-sharing-photos {
-  background-image: url(~/assets/img/sked-sharing-photos.png);
+  background-image: url(~assets/img/sked-sharing-photos.png);
 }
 .sked-button:hover,
 .sked-dialog:hover,
