@@ -1,24 +1,20 @@
 <template>
   <div class="wrapper">
     <q-header-small class-name="evan-dennis">
-      <h1 v-html="$t('faq.title')"/>
+      <h1 v-html="$t('faq.title')" />
     </q-header-small>
     <div class="section">
-      <div
-        v-for="group in $t('faq.groups')"
-        :key="group.title"
-        class="row">
+      <div v-for="group in $t('faq.groups')" :key="group.title" class="row">
         <div class="col-md-10 ml-auto mr-auto">
-          <h2
-            class="title"
-            v-html="group.title"/>
+          <h2 class="title" v-html="group.title" />
           <br>
           <div class="row justify-content-center">
             <q-collapse
-              v-for="(faq,index) in group.items"
+              v-for="(faq, index) in group.items"
               :key="faq.question"
-              :title="`${index+1} - ${faq.question}`"
-              :desc="faq.answer" />
+              :title="`${index + 1} - ${faq.question}`"
+              :desc="faq.answer"
+            />
           </div>
         </div>
       </div>
@@ -40,6 +36,6 @@ export default {
 </script>
 <style>
 .evan-dennis {
-  background-image: url(~/assets/img/evan-dennis-75563-unsplash.jpg);
+  background-image: url(~assets/img/evan-dennis-75563-unsplash.jpg);
 }
 </style>
