@@ -1,10 +1,9 @@
 <template>
   <div class="wrapper">
     <q-header-small class-name="privacy-policy">
-      <h1 v-html="$t('policy.title')"/>
+      <h1 v-html="$t('policy.title')" />
     </q-header-small>
     <div class="main">
-
       <div class="section">
         <div class="container">
           <div class="row">
@@ -12,56 +11,99 @@
               <h2>Privacy policy</h2>
 
               <p>Effective date: October 16, 2018</p>
-              <p>Skedr.io ("us", "we", or "our") operates the https://skedr.io website and the Skedr.io mobile application (the "Service").</p>
+              <p>
+                Skedr.io ("us", "we", or "our") operates the https://skedr.io website and the Skedr.io mobile
+                application (the "Service").
+              </p>
 
-              <p>This page informs you of our policies regarding the collection, use, and disclosure of personal data when you use our Service and the choices you have associated with that data. Our Privacy Policy  for Skedr.io is managed through <a href="https://www.freeprivacypolicy.com/free-privacy-policy-generator.php">Free Privacy Policy</a>.</p>
+              <p>
+                This page informs you of our policies regarding the collection, use, and disclosure of personal data
+                when you use our Service and the choices you have associated with that data. Our Privacy Policy for
+                Skedr.io is managed through
+                <a href="https://www.freeprivacypolicy.com/free-privacy-policy-generator.php">Free Privacy Policy</a>.
+              </p>
 
-              <p>We use your data to provide and improve the Service. By using the Service, you agree to the collection and use of information in accordance with this policy. Unless otherwise defined in this Privacy Policy, terms used in this Privacy Policy have the same meanings as in our Terms and Conditions.</p>
+              <p>
+                We use your data to provide and improve the Service. By using the Service, you agree to the collection
+                and use of information in accordance with this policy. Unless otherwise defined in this Privacy Policy,
+                terms used in this Privacy Policy have the same meanings as in our Terms and Conditions.
+              </p>
             </div>
           </div>
         </div>
       </div>
 
-
       <div class="section">
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Information Collection And Use</h2>
 
-              <p>We collect several different types of information for various purposes to provide and improve our Service to you.</p>
+              <p>
+                We collect several different types of information for various purposes to provide and improve our
+                Service to you.
+              </p>
 
               <h3>Types of Data Collected</h3>
 
               <h4>Personal Data</h4>
 
-              <p>While using our Service, we may ask you to provide us with certain personally identifiable information that can be used to contact or identify you ("Personal Data"). Personally identifiable information may include, but is not limited to:</p>
+              <p>
+                While using our Service, we may ask you to provide us with certain personally identifiable information
+                that can be used to contact or identify you ("Personal Data"). Personally identifiable information may
+                include, but is not limited to:
+              </p>
 
               <ul>
-                <li>Email address</li><li>First name and last name</li><li>Cookies and Usage Data</li>
+                <li>Email address</li>
+                <li>First name and last name</li>
+                <li>Cookies and Usage Data</li>
               </ul>
               <h4>Usage Data</h4>
 
-              <p>We may also collect information that your browser sends whenever you visit our Service or when you access the Service by or through a mobile device ("Usage Data").</p>
-              <p>This Usage Data may include information such as your computer's Internet Protocol address (e.g. IP address), browser type, browser version, the pages of our Service that you visit, the time and date of your visit, the time spent on those pages, unique device identifiers and other diagnostic data.</p>
-              <p>When you access the Service by or through a mobile device, this Usage Data may include information such as the type of mobile device you use, your mobile device unique ID, the IP address of your mobile device, your mobile operating system, the type of mobile Internet browser you use, unique device identifiers and other diagnostic data.</p>
+              <p>
+                We may also collect information that your browser sends whenever you visit our Service or when you
+                access the Service by or through a mobile device ("Usage Data").
+              </p>
+              <p>
+                This Usage Data may include information such as your computer's Internet Protocol address (e.g. IP
+                address), browser type, browser version, the pages of our Service that you visit, the time and date of
+                your visit, the time spent on those pages, unique device identifiers and other diagnostic data.
+              </p>
+              <p>
+                When you access the Service by or through a mobile device, this Usage Data may include information such
+                as the type of mobile device you use, your mobile device unique ID, the IP address of your mobile
+                device, your mobile operating system, the type of mobile Internet browser you use, unique device
+                identifiers and other diagnostic data.
+              </p>
 
               <h4>Tracking & Cookies Data</h4>
-              <p>We use cookies and similar tracking technologies to track the activity on our Service and hold certain information.</p>
-              <p>Cookies are files with small amount of data which may include an anonymous unique identifier. Cookies are sent to your browser from a website and stored on your device. Tracking technologies also used are beacons, tags, and scripts to collect and track information and to improve and analyze our Service.</p>
-              <p>You can instruct your browser to refuse all cookies or to indicate when a cookie is being sent. However, if you do not accept cookies, you may not be able to use some portions of our Service.</p>
+              <p>
+                We use cookies and similar tracking technologies to track the activity on our Service and hold certain
+                information.
+              </p>
+              <p>
+                Cookies are files with small amount of data which may include an anonymous unique identifier. Cookies
+                are sent to your browser from a website and stored on your device. Tracking technologies also used are
+                beacons, tags, and scripts to collect and track information and to improve and analyze our Service.
+              </p>
+              <p>
+                You can instruct your browser to refuse all cookies or to indicate when a cookie is being sent. However,
+                if you do not accept cookies, you may not be able to use some portions of our Service.
+              </p>
               <p>Examples of Cookies we use:</p>
               <ul>
                 <li><strong>Session Cookies.</strong> We use Session Cookies to operate our Service.</li>
-                <li><strong>Preference Cookies.</strong> We use Preference Cookies to remember your preferences and various settings.</li>
+                <li>
+                  <strong>Preference Cookies.</strong> We use Preference Cookies to remember your preferences and
+                  various settings.
+                </li>
                 <li><strong>Security Cookies.</strong> We use Security Cookies for security purposes.</li>
               </ul>
             </div>
           </div>
         </div>
       </div>
-
 
       <div class="section">
         <div class="container">
@@ -88,10 +130,25 @@
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
               <h2>Transfer Of Data</h2>
-              <p>Your information, including Personal Data, may be transferred to — and maintained on — computers located outside of your state, province, country or other governmental jurisdiction where the data protection laws may differ than those from your jurisdiction.</p>
-              <p>If you are located outside Spain and choose to provide information to us, please note that we transfer the data, including Personal Data, to Spain and process it there.</p>
-              <p>Your consent to this Privacy Policy followed by your submission of such information represents your agreement to that transfer.</p>
-              <p>Skedr.io will take all steps reasonably necessary to ensure that your data is treated securely and in accordance with this Privacy Policy and no transfer of your Personal Data will take place to an organization or a country unless there are adequate controls in place including the security of your data and other personal information.</p>
+              <p>
+                Your information, including Personal Data, may be transferred to — and maintained on — computers located
+                outside of your state, province, country or other governmental jurisdiction where the data protection
+                laws may differ than those from your jurisdiction.
+              </p>
+              <p>
+                If you are located outside Spain and choose to provide information to us, please note that we transfer
+                the data, including Personal Data, to Spain and process it there.
+              </p>
+              <p>
+                Your consent to this Privacy Policy followed by your submission of such information represents your
+                agreement to that transfer.
+              </p>
+              <p>
+                Skedr.io will take all steps reasonably necessary to ensure that your data is treated securely and in
+                accordance with this Privacy Policy and no transfer of your Personal Data will take place to an
+                organization or a country unless there are adequate controls in place including the security of your
+                data and other personal information.
+              </p>
             </div>
           </div>
         </div>
@@ -101,7 +158,6 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Disclosure Of Data</h2>
 
               <h3>Legal Requirements</h3>
@@ -121,21 +177,40 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Security Of Data</h2>
-              <p>The security of your data is important to us, but remember that no method of transmission over the Internet, or method of electronic storage is 100% secure. While we strive to use commercially acceptable means to protect your Personal Data, we cannot guarantee its absolute security.</p>
+              <p>
+                The security of your data is important to us, but remember that no method of transmission over the
+                Internet, or method of electronic storage is 100% secure. While we strive to use commercially acceptable
+                means to protect your Personal Data, we cannot guarantee its absolute security.
+              </p>
 
               <h2>Service Providers</h2>
-              <p>We may employ third party companies and individuals to facilitate our Service ("Service Providers"), to provide the Service on our behalf, to perform Service-related services or to assist us in analyzing how our Service is used.</p>
-              <p>These third parties have access to your Personal Data only to perform these tasks on our behalf and are obligated not to disclose or use it for any other purpose.</p>
+              <p>
+                We may employ third party companies and individuals to facilitate our Service ("Service Providers"), to
+                provide the Service on our behalf, to perform Service-related services or to assist us in analyzing how
+                our Service is used.
+              </p>
+              <p>
+                These third parties have access to your Personal Data only to perform these tasks on our behalf and are
+                obligated not to disclose or use it for any other purpose.
+              </p>
 
               <h3>Analytics</h3>
               <p>We may use third-party Service Providers to monitor and analyze the use of our Service.</p>
               <ul>
                 <li>
                   <p><strong>Google Analytics</strong></p>
-                  <p>Google Analytics is a web analytics service offered by Google that tracks and reports website traffic. Google uses the data collected to track and monitor the use of our Service. This data is shared with other Google services. Google may use the collected data to contextualize and personalize the ads of its own advertising network.</p>
-                  <p>For more information on the privacy practices of Google, please visit the Google Privacy & Terms web page: <a href="https://policies.google.com/privacy?hl=en">https://policies.google.com/privacy?hl=en</a></p>
+                  <p>
+                    Google Analytics is a web analytics service offered by Google that tracks and reports website
+                    traffic. Google uses the data collected to track and monitor the use of our Service. This data is
+                    shared with other Google services. Google may use the collected data to contextualize and
+                    personalize the ads of its own advertising network.
+                  </p>
+                  <p>
+                    For more information on the privacy practices of Google, please visit the Google Privacy & Terms web
+                    page:
+                    <a href="https://policies.google.com/privacy?hl=en">https://policies.google.com/privacy?hl=en</a>
+                  </p>
                 </li>
               </ul>
             </div>
@@ -143,45 +218,59 @@
         </div>
       </div>
 
-
-
       <div class="section">
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Links To Other Sites</h2>
-              <p>Our Service may contain links to other sites that are not operated by us. If you click on a third party link, you will be directed to that third party's site. We strongly advise you to review the Privacy Policy of every site you visit.</p>
-              <p>We have no control over and assume no responsibility for the content, privacy policies or practices of any third party sites or services.</p>
+              <p>
+                Our Service may contain links to other sites that are not operated by us. If you click on a third party
+                link, you will be directed to that third party's site. We strongly advise you to review the Privacy
+                Policy of every site you visit.
+              </p>
+              <p>
+                We have no control over and assume no responsibility for the content, privacy policies or practices of
+                any third party sites or services.
+              </p>
             </div>
           </div>
         </div>
       </div>
 
-
       <div class="section">
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Children's Privacy</h2>
               <p>Our Service does not address anyone under the age of 18 ("Children").</p>
-              <p>We do not knowingly collect personally identifiable information from anyone under the age of 18. If you are a parent or guardian and you are aware that your Children has provided us with Personal Data, please contact us. If we become aware that we have collected Personal Data from children without verification of parental consent, we take steps to remove that information from our servers.</p>
+              <p>
+                We do not knowingly collect personally identifiable information from anyone under the age of 18. If you
+                are a parent or guardian and you are aware that your Children has provided us with Personal Data, please
+                contact us. If we become aware that we have collected Personal Data from children without verification
+                of parental consent, we take steps to remove that information from our servers.
+              </p>
             </div>
           </div>
         </div>
       </div>
 
-
       <div class="section">
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Changes To This Privacy Policy</h2>
-              <p>We may update our Privacy Policy from time to time. We will notify you of any changes by posting the new Privacy Policy on this page.</p>
-              <p>We will let you know via email and/or a prominent notice on our Service, prior to the change becoming effective and update the "effective date" at the top of this Privacy Policy.</p>
-              <p>You are advised to review this Privacy Policy periodically for any changes. Changes to this Privacy Policy are effective when they are posted on this page.</p>
+              <p>
+                We may update our Privacy Policy from time to time. We will notify you of any changes by posting the new
+                Privacy Policy on this page.
+              </p>
+              <p>
+                We will let you know via email and/or a prominent notice on our Service, prior to the change becoming
+                effective and update the "effective date" at the top of this Privacy Policy.
+              </p>
+              <p>
+                You are advised to review this Privacy Policy periodically for any changes. Changes to this Privacy
+                Policy are effective when they are posted on this page.
+              </p>
             </div>
           </div>
         </div>
@@ -191,7 +280,6 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Contact Us</h2>
               <p>If you have any questions about this Privacy Policy, please contact us:</p>
               <ul>
@@ -201,7 +289,8 @@
           </div>
         </div>
       </div>
-  </div> </div>
+    </div>
+  </div>
 </template>
 <script>
 import QHeaderSmall from '~/components/QHeaderSmall'
