@@ -1,10 +1,9 @@
 <template>
   <div class="wrapper">
     <q-header-small class-name="privacy-policy">
-      <h1 v-html="$t('terms.title')"/>
+      <h1 v-html="$t('terms.title')" />
     </q-header-small>
     <div class="main">
-
       <div class="section">
         <div class="container">
           <div class="row">
@@ -12,10 +11,19 @@
               <h1>Terms and Conditions</h1>
 
               <p>Last updated: October 16, 2018</p>
-              <p>These Terms and Conditions ("Terms", "Terms and Conditions") govern your relationship with https://skedr.io website (the "Service") operated by Skedr.io ("us", "we", or "our").</p>
+              <p>
+                These Terms and Conditions ("Terms", "Terms and Conditions") govern your relationship with
+                https://skedr.io website (the "Service") operated by Skedr.io ("us", "we", or "our").
+              </p>
               <p>Please read these Terms and Conditions carefully before using the Service.</p>
-              <p>Your access to and use of the Service is conditioned on your acceptance of and compliance with these Terms. These Terms apply to all visitors, users and others who access or use the Service.</p>
-              <p>By accessing or using the Service you agree to be bound by these Terms. If you disagree with any part of the terms then you may not access the Service.</p>
+              <p>
+                Your access to and use of the Service is conditioned on your acceptance of and compliance with these
+                Terms. These Terms apply to all visitors, users and others who access or use the Service.
+              </p>
+              <p>
+                By accessing or using the Service you agree to be bound by these Terms. If you disagree with any part of
+                the terms then you may not access the Service.
+              </p>
             </div>
           </div>
         </div>
@@ -25,16 +33,33 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Subscriptions</h2>
 
-              <p>Some parts of the Service are billed on a subscription basis ("Subscription(s)"). You will be billed in advance on a recurring and periodic basis ("Billing Cycle"). Billing cycles are set either on a monthly or annual basis, depending on the type of subscription plan you select when purchasing a Subscription.</p>
+              <p>
+                Some parts of the Service are billed on a subscription basis ("Subscription(s)"). You will be billed in
+                advance on a recurring and periodic basis ("Billing Cycle"). Billing cycles are set either on a monthly
+                or annual basis, depending on the type of subscription plan you select when purchasing a Subscription.
+              </p>
 
-              <p>At the end of each Billing Cycle, your Subscription will automatically renew under the exact same conditions unless you cancel it or Skedr.io cancels it. You may cancel your Subscription renewal either through your online account management page or by contacting Skedr.io customer support team.</p>
+              <p>
+                At the end of each Billing Cycle, your Subscription will automatically renew under the exact same
+                conditions unless you cancel it or Skedr.io cancels it. You may cancel your Subscription renewal either
+                through your online account management page or by contacting Skedr.io customer support team.
+              </p>
 
-              <p>A valid payment method, including credit card or PayPal, is required to process the payment for your Subscription. You shall provide Skedr.io with accurate and complete billing information including full name, address, state, zip code, telephone number, and a valid payment method information. By submitting such payment information, you automatically authorize Skedr.io to charge all Subscription fees incurred through your account to any such payment instruments.</p>
+              <p>
+                A valid payment method, including credit card or PayPal, is required to process the payment for your
+                Subscription. You shall provide Skedr.io with accurate and complete billing information including full
+                name, address, state, zip code, telephone number, and a valid payment method information. By submitting
+                such payment information, you automatically authorize Skedr.io to charge all Subscription fees incurred
+                through your account to any such payment instruments.
+              </p>
 
-              <p>Should automatic billing fail to occur for any reason, Skedr.io will issue an electronic invoice indicating that you must proceed manually, within a certain deadline date, with the full payment corresponding to the billing period as indicated on the invoice.</p>
+              <p>
+                Should automatic billing fail to occur for any reason, Skedr.io will issue an electronic invoice
+                indicating that you must proceed manually, within a certain deadline date, with the full payment
+                corresponding to the billing period as indicated on the invoice.
+              </p>
             </div>
           </div>
         </div>
@@ -44,14 +69,23 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Fee Changes</h2>
 
-              <p>Skedr.io, in its sole discretion and at any time, may modify the Subscription fees for the Subscriptions. Any Subscription fee change will become effective at the end of the then-current Billing Cycle.</p>
+              <p>
+                Skedr.io, in its sole discretion and at any time, may modify the Subscription fees for the
+                Subscriptions. Any Subscription fee change will become effective at the end of the then-current Billing
+                Cycle.
+              </p>
 
-              <p>Skedr.io will provide you with a reasonable prior notice of any change in Subscription fees to give you an opportunity to terminate your Subscription before such change becomes effective.</p>
+              <p>
+                Skedr.io will provide you with a reasonable prior notice of any change in Subscription fees to give you
+                an opportunity to terminate your Subscription before such change becomes effective.
+              </p>
 
-              <p>Your continued use of the Service after the Subscription fee change comes into effect constitutes your agreement to pay the modified Subscription fee amount.</p>
+              <p>
+                Your continued use of the Service after the Subscription fee change comes into effect constitutes your
+                agreement to pay the modified Subscription fee amount.
+              </p>
             </div>
           </div>
         </div>
@@ -61,10 +95,12 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Refunds</h2>
 
-              <p>Certain refund requests for Subscriptions may be considered by Skedr.io on a case-by-case basis and granted in sole discretion of Skedr.io.</p>
+              <p>
+                Certain refund requests for Subscriptions may be considered by Skedr.io on a case-by-case basis and
+                granted in sole discretion of Skedr.io.
+              </p>
             </div>
           </div>
         </div>
@@ -74,14 +110,24 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Accounts</h2>
 
-              <p>When you create an account with us, you must provide us information that is accurate, complete, and current at all times. Failure to do so constitutes a breach of the Terms, which may result in immediate termination of your account on our Service.</p>
+              <p>
+                When you create an account with us, you must provide us information that is accurate, complete, and
+                current at all times. Failure to do so constitutes a breach of the Terms, which may result in immediate
+                termination of your account on our Service.
+              </p>
 
-              <p>You are responsible for safeguarding the password that you use to access the Service and for any activities or actions under your password, whether your password is with our Service or a third-party service.</p>
+              <p>
+                You are responsible for safeguarding the password that you use to access the Service and for any
+                activities or actions under your password, whether your password is with our Service or a third-party
+                service.
+              </p>
 
-              <p>You agree not to disclose your password to any third party. You must notify us immediately upon becoming aware of any breach of security or unauthorized use of your account.</p>
+              <p>
+                You agree not to disclose your password to any third party. You must notify us immediately upon becoming
+                aware of any breach of security or unauthorized use of your account.
+              </p>
             </div>
           </div>
         </div>
@@ -91,10 +137,14 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Intellectual Property</h2>
 
-              <p>The Service and its original content, features and functionality are and will remain the exclusive property of Skedr.io and its licensors. The Service is protected by copyright, trademark, and other laws of both the Spain and foreign countries. Our trademarks and trade dress may not be used in connection with any product or service without the prior written consent of Skedr.io.</p>
+              <p>
+                The Service and its original content, features and functionality are and will remain the exclusive
+                property of Skedr.io and its licensors. The Service is protected by copyright, trademark, and other laws
+                of both the Spain and foreign countries. Our trademarks and trade dress may not be used in connection
+                with any product or service without the prior written consent of Skedr.io.
+              </p>
             </div>
           </div>
         </div>
@@ -105,9 +155,21 @@
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
               <h2>Links To Other Web Sites</h2>
-              <p>Our Service may contain links to third-party web sites or services that are not owned or controlled by Skedr.io.</p>
-              <p>Skedr.io has no control over, and assumes no responsibility for, the content, privacy policies, or practices of any third party web sites or services. You further acknowledge and agree that Skedr.io shall not be responsible or liable, directly or indirectly, for any damage or loss caused or alleged to be caused by or in connection with use of or reliance on any such content, goods or services available on or through any such web sites or services.</p>
-              <p>We strongly advise you to read the terms and conditions and privacy policies of any third-party web sites or services that you visit.</p>
+              <p>
+                Our Service may contain links to third-party web sites or services that are not owned or controlled by
+                Skedr.io.
+              </p>
+              <p>
+                Skedr.io has no control over, and assumes no responsibility for, the content, privacy policies, or
+                practices of any third party web sites or services. You further acknowledge and agree that Skedr.io
+                shall not be responsible or liable, directly or indirectly, for any damage or loss caused or alleged to
+                be caused by or in connection with use of or reliance on any such content, goods or services available
+                on or through any such web sites or services.
+              </p>
+              <p>
+                We strongly advise you to read the terms and conditions and privacy policies of any third-party web
+                sites or services that you visit.
+              </p>
             </div>
           </div>
         </div>
@@ -118,8 +180,14 @@
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
               <h2>Termination</h2>
-              <p>We may terminate or suspend your account immediately, without prior notice or liability, for any reason whatsoever, including without limitation if you breach the Terms.</p>
-              <p>Upon termination, your right to use the Service will immediately cease. If you wish to terminate your account, you may simply discontinue using the Service.</p>
+              <p>
+                We may terminate or suspend your account immediately, without prior notice or liability, for any reason
+                whatsoever, including without limitation if you breach the Terms.
+              </p>
+              <p>
+                Upon termination, your right to use the Service will immediately cease. If you wish to terminate your
+                account, you may simply discontinue using the Service.
+              </p>
             </div>
           </div>
         </div>
@@ -130,7 +198,16 @@
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
               <h2>Limitation Of Liability</h2>
-              <p>In no event shall Skedr.io, nor its directors, employees, partners, agents, suppliers, or affiliates, be liable for any indirect, incidental, special, consequential or punitive damages, including without limitation, loss of profits, data, use, goodwill, or other intangible losses, resulting from (i) your access to or use of or inability to access or use the Service; (ii) any conduct or content of any third party on the Service; (iii) any content obtained from the Service; and (iv) unauthorized access, use or alteration of your transmissions or content, whether based on warranty, contract, tort (including negligence) or any other legal theory, whether or not we have been informed of the possibility of such damage, and even if a remedy set forth herein is found to have failed of its essential purpose.</p>
+              <p>
+                In no event shall Skedr.io, nor its directors, employees, partners, agents, suppliers, or affiliates, be
+                liable for any indirect, incidental, special, consequential or punitive damages, including without
+                limitation, loss of profits, data, use, goodwill, or other intangible losses, resulting from (i) your
+                access to or use of or inability to access or use the Service; (ii) any conduct or content of any third
+                party on the Service; (iii) any content obtained from the Service; and (iv) unauthorized access, use or
+                alteration of your transmissions or content, whether based on warranty, contract, tort (including
+                negligence) or any other legal theory, whether or not we have been informed of the possibility of such
+                damage, and even if a remedy set forth herein is found to have failed of its essential purpose.
+              </p>
             </div>
           </div>
         </div>
@@ -141,8 +218,18 @@
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
               <h2>Disclaimer</h2>
-              <p>Your use of the Service is at your sole risk. The Service is provided on an "AS IS" and "AS AVAILABLE" basis. The Service is provided without warranties of any kind, whether express or implied, including, but not limited to, implied warranties of merchantability, fitness for a particular purpose, non-infringement or course of performance.</p>
-              <p>Skedr.io its subsidiaries, affiliates, and its licensors do not warrant that a) the Service will function uninterrupted, secure or available at any particular time or location; b) any errors or defects will be corrected; c) the Service is free of viruses or other harmful components; or d) the results of using the Service will meet your requirements.</p>
+              <p>
+                Your use of the Service is at your sole risk. The Service is provided on an "AS IS" and "AS AVAILABLE"
+                basis. The Service is provided without warranties of any kind, whether express or implied, including,
+                but not limited to, implied warranties of merchantability, fitness for a particular purpose,
+                non-infringement or course of performance.
+              </p>
+              <p>
+                Skedr.io its subsidiaries, affiliates, and its licensors do not warrant that a) the Service will
+                function uninterrupted, secure or available at any particular time or location; b) any errors or defects
+                will be corrected; c) the Service is free of viruses or other harmful components; or d) the results of
+                using the Service will meet your requirements.
+              </p>
             </div>
           </div>
         </div>
@@ -153,8 +240,17 @@
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
               <h2>Governing Law</h2>
-              <p>These Terms shall be governed and construed in accordance with the laws of Spain, without regard to its conflict of law provisions.</p>
-              <p>Our failure to enforce any right or provision of these Terms will not be considered a waiver of those rights. If any provision of these Terms is held to be invalid or unenforceable by a court, the remaining provisions of these Terms will remain in effect. These Terms constitute the entire agreement between us regarding our Service, and supersede and replace any prior agreements we might have between us regarding the Service.</p>
+              <p>
+                These Terms shall be governed and construed in accordance with the laws of Spain, without regard to its
+                conflict of law provisions.
+              </p>
+              <p>
+                Our failure to enforce any right or provision of these Terms will not be considered a waiver of those
+                rights. If any provision of these Terms is held to be invalid or unenforceable by a court, the remaining
+                provisions of these Terms will remain in effect. These Terms constitute the entire agreement between us
+                regarding our Service, and supersede and replace any prior agreements we might have between us regarding
+                the Service.
+              </p>
             </div>
           </div>
         </div>
@@ -164,12 +260,18 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Changes</h2>
 
-              <p>We reserve the right, at our sole discretion, to modify or replace these Terms at any time. If a revision is material we will try to provide at least 30 days notice prior to any new terms taking effect. What constitutes a material change will be determined at our sole discretion.</p>
+              <p>
+                We reserve the right, at our sole discretion, to modify or replace these Terms at any time. If a
+                revision is material we will try to provide at least 30 days notice prior to any new terms taking
+                effect. What constitutes a material change will be determined at our sole discretion.
+              </p>
 
-              <p>By continuing to access or use our Service after those revisions become effective, you agree to be bound by the revised terms. If you do not agree to the new terms, please stop using the Service.</p>
+              <p>
+                By continuing to access or use our Service after those revisions become effective, you agree to be bound
+                by the revised terms. If you do not agree to the new terms, please stop using the Service.
+              </p>
             </div>
           </div>
         </div>
@@ -179,7 +281,6 @@
         <div class="container">
           <div class="row">
             <div class="col-md-8 ml-auto mr-auto">
-
               <h2>Contact Us</h2>
 
               <p>If you have any questions about these Terms, please contact us.</p>
